+++ pytest_rerunfailures.py
@@ -12,7 +12,6 @@ from contextlib import suppress
 
 import pytest
 from _pytest.outcomes import fail
-from _pytest.python import Function
 from _pytest.runner import runtestprotocol
 from packaging.version import parse as parse_version
 
@@ -525,16 +524,17 @@ def pytest_runtest_teardown(item, nextitem):
         # clean cashed results from any level of setups
         _remove_cached_results_from_failed_fixtures(item)
 
-        if PYTEST_GTE_63:
-            for key in list(item.session._setupstate.stack.keys()):
-                if type(key) != Function:
-                    del item.session._setupstate.stack[key]
-        else:
-            for node in list(item.session._setupstate.stack):
-                if type(node) != Function:
-                    item.session._setupstate.stack.remove(node)
+        if item in item.session._setupstate.stack:
+            if PYTEST_GTE_63:
+                for key in list(item.session._setupstate.stack.keys()):
+                    if key != item:
+                        del item.session._setupstate.stack[key]
+            else:
+                for node in list(item.session._setupstate.stack):
+                    if node != item:
+                        item.session._setupstate.stack.remove(node)
 
-        item.teardown()
+            item.teardown()
 
 
 @pytest.hookimpl(hookwrapper=True)