+++ CHANGES.rst
@@ -1,7 +1,16 @@
 Changelog
 =========
 
-11.1 (unreleased)
+11.1.1 (unreleased)
+-------------------
+
+Bug fixes
++++++++++
+
+- Fix crash during teardown when runtest protocol hook is overwritten by another plugin.
+- Fix crash during teardown when TestCase class is used as base class.
+
+11.1 (2023-02-09)
 -----------------
 
 Bug fixes